Learns CC_TX transmit period per CC_TX #9

+++ Manager.h
@@ -40,7 +40,7 @@ private:
      * CC TRX (e.g. EDF IAMs)                *
      *****************************************/
     static const uint8_t  MAX_CC_TRXS = 35;   /* number of CC TRXs (e.g. EDF IAMs) */
-	static const unsigned long CC_TRX_TIMEOUT = 100; // milliseconds to wait for reply
+	static const uint32_t CC_TRX_TIMEOUT = 100; // milliseconds to wait for reply
     uint32_t cc_trx_ids[MAX_CC_TRXS];
     uint8_t  i_next_cc_trx; // index into cc_trx_ids
 	uint32_t id_next_cc_trx;// actual id (i.e. a copy of a cc_trx_ids[i_next_cc_trx])
@@ -49,7 +49,7 @@ private:
 	uint8_t retries; // for polling CC TRXs
 	static const uint8_t MAX_RETRIES = 5; // for polling CC TRXs
 
-	unsigned long timecode_polled_first_cc_trx;
+	uint32_t timecode_polled_first_cc_trx;
 
 	/***************************
 	 * Private methods
